Abstract HLT path finding etc to a new HLTTauDQMPath class
